Improve how conversations work and make it easier to handle them
With this commit, we do something new: update the interactions in the prompt
after each message. This is necessary beacuse ollama stores the conversation as
a vector. If we ever integrate with the command-line "llm"
package (https://llm.datasette.io/en/stable/index.html), it works in a similar
way. However, Open AI and Google Cloud Vertex rely on a list of interactions.
Because of this discrepancy, we store data in the interactions list, which is
now opaque, and must be added to with the new function
`llm-chat-prompt-append-response'.

The README has an example of how this all works.