Fix one cause of missing tokens when streaming for ollama provider
One issue is that if there was content streamed that was incomplete JSON, we
would never parse the incomplete part.  Now we make sure to only advance when we
successfully parse, and try to be more precise about getting only valid JSON.

This does not completely solve the problem, however.  The other causes of
missing content are currently unknown.

This is a partial fix for https://github.com/s-kostyaev/ellama/issues/8.