Add plz and media type and event source extensions
This adds plz to the LLM library.

*plz.el* is a slightly modified version of upstream.

1. It adds a slot for the process object to the response struct, so we
can access properties of the process in various places. In our
discussion about the plz streaming PR, @alphapapa suggested this at
some point.

2. The plz function has an additional option called :process-filter
that allows setting the process filter when the curl process is
created with make-process. This is needed to set the process filter in
the synchronous and asynchronous cases. In the asynchronous case you
could to this immediately after the process is created. In the
synchronous case you can't get at the process object to install a
filter.

*plz-media-type.el* contains code that can be used implement and
customize response decoding for various media types. It ships with
default media types for application/json, application/html,
application/xml and the default application/octet-stream. A media type
can support "normal" and "straming" formats.

*plz-event-source.el* contains a media type implementation for
text/event-stream, aka server sent events, and an implementation of an
event source class and parser according to the HTML living standard.