Strip plz changes and add JSON array stream media type
- This removes the process slot from the plz response struct which I
added initially. We don't need it when using the :as 'buffer option
when calling out to plz. We now only depend on one additional change
in plz, which is setting the process filter via an option.

- This adds the plz-media-type:application/json-array media type which
can be used to stream a JSON array as used by Vertex/Gemini. It's a
poor mans streaming JSON parser. You can register a handler that gets
called for each object in the outer most JSON array of the
response. You can use it in a similar way as the event stream handler.

```
(let* ((api-key "my-key")
       (project "my-project"))
  (plz-media-type-request
    'post (format
           (concat "https://us-central1-aiplatform.googleapis.com"
                   "/v1/projects/%s/locations"
                   "/us-central1/publishers/google/models"
                   "/gemini-1.0-pro:streamGenerateContent")
           project)
    :as `(media-types (("application/json"
                        . ,(plz-media-type:application/json-array
                            :handler (lambda (object)
                                       (message "Object: %s" object))))))
    :body (json-encode
           `((contents .
                       [((role . "user")
                         (parts .
                                [((text . "Hello"))]))])
             (generation_config
              (maxOutputTokens . 2048)
              (temperature . 1)
              (topP . 0.4))
             (safetySettings .
                             [((category . "HARM_CATEGORY_HATE_SPEECH")
                               (threshold . "BLOCK_MEDIUM_AND_ABOVE"))
                              ((category . "HARM_CATEGORY_DANGEROUS_CONTENT")
                               (threshold . "BLOCK_MEDIUM_AND_ABOVE"))
                              ((category . "HARM_CATEGORY_SEXUALLY_EXPLICIT")
                               (threshold . "BLOCK_MEDIUM_AND_ABOVE"))
                              ((category . "HARM_CATEGORY_HARASSMENT")
                               (threshold . "BLOCK_MEDIUM_AND_ABOVE"))])))
    :headers `(("Authorization" . ,(format "Bearer %s" api-key))
               ("Content-Type" . "application/json"))
    :then (lambda (response)
            (message "Done."))))
```

See the tests in [1] for more examples.

Unfortunatly with Gemini/Vertex wanting a streaming parser for
application/json we need to distinguish now 2 cases:

- Handle application/json in streaming way
- Handle application/json in non-streaming way

In order to use this, I would suggest to pass the media types via a
:media-types option to the following functions:

- llm-request-plz-event-stream
- llm-request-plz-sync
- llm-request-plz-async

We could use the plz-media-types as the default, but for any media
type that needs to register event handlers for a streaming media type,
I think it's best if we wire the list of media types (and the
callbacks) together in each provider.

Wdyt?

[1] https://github.com/r0man/plz.el/blob/plz-media-type/tests/test-plz-media-type.el#L227-L256