Fix error handling.
The on-error callback should be called with 2 arguments, a code and
the payload. There are 2 cases:

HTTP errors do have a status code and response. In this case we call
the error handler with the HTTP status code and the HTTP body.

Curl error don't have a response. These are errors like DNS errors,
connection timouts etc. In those cases we call the error handler with
the curl exit code, and an error message (derived from that code).