Fix error handling in Gemini and Vertex provider
The :else callback of the json-array media type doesn't contain the
error messages. The media type always streams back the objects in the
JSON array via the :handlers. In case of errors, the Vertex and
Gemini (I think, I don't have access to it) response is also an array,
with the elements being the error objects.