Widen before parsing the response
I sometimes see this error:

```
Debugger entered--Lisp error: (wrong-type-argument number-or-marker-p nil)
  <=(200 nil 299)
  (let* ((status val)) (<= 200 status 299))
```

There seem to be some workarounds in plz about making sure
plz--respond is called. I believe it can happen in some cases that it
gets called with a narrowed buffer. In that case this error
occurs. The call to widen seems to fix the issue.