Fix issue where gzip interfered with streaming
This usually occured in the first attempt at streaming.  The issue is that there
really is no good way to turn gzip encoding off, which has already been
noted (https://lists.gnu.org/r/emacs-devel/2016-03/msg01533.html).  Turning it
off globally for everyone, at least while we stream, is the best we can do.
This may cause problems for any package that needs to ensure the gzip is in fact
on.