Add support for HTTP/HTTPS proxies (#49)
This adds support for HTTP/HTTPS proxies. The code that was dealing
with this in plz.el wasn't migrated and we did not had any tests for
this. This should now be fixed. See:

- https://github.com/r0man/plz-media-type/pull/8
- https://github.com/r0man/plz-event-source/pull/5

It fixes the issue #48 reported by @theasp. He tested the code with
Ollama behind a Squid proxy. I tested it with Vertex and OpenAI behind
a tinyproxy.

https://github.com/ahyatt/llm/issues/48