Add llm-prompt, a module for constructing and storing prompts
llm-prompt is a way to define prompts in applications with templatized variables
that can either be simple variables substituted in at the LLM call, or lists or
generators which allows different contextual streams to fill up the context
window to the desired size.

Because it uses token counting so much, we are getting rid of the Gemini &
Vertex token counting call, which would be too expensive.  It may be added back
in the future in some other way.