Unvendor plz-event-source and plz-media-type (#87)
This PR aims to unvendor plz-event-source and plz-media-type. According
to <https://github.com/ahyatt/llm/pull/46#issuecomment-2387962790>, this
should not cause any problems.

I chose to define `0.1.0` as the minimum required versions, because the
vendored packages are seemingly older, and it seems like established
Elisp practice is to specify the "lowest version that is expected to
reasonably function".