Handle plz-error-message error type (#95)
In some cases, plz may return a `plz-error` that does not contain a
response or curl error, but just an error message. I think this should
be passed to the `:on-error` callback.

On a related note, I've noticed that I don't completely understand the
error handling model of llm. Some providers seem to have a
`llm-provider-chat-extract-error` function which expects structured
errors and is also called in the `:on-error` callback inside of
`(llm-chat-async 'llm-standard-chat-provider ...)`. But looking at the
implementation of `llm-request-plz-async`, this callback will only ever
be called with stringly data.

Maybe `llm-provider-chat-extract-error` should only be called in the
`:on-success` callback? That would make more sense, if I correctly
understand the code.