Add multimodal support for openai and gemini (#88)
I've added multimodal support for two of the providers (openai &
vertex/gemini) so images can be included as part of the input prompt.

I settled on the following system for doing this:
1. I follow the convention that if the role is user and content is a
list, then it is a multipart message.
2. If it's a string everything works the same as before, and if the role
is assistant and content is a list, then I assume it is a list of
function calls and use the existing code.
3. If content is a list and role is user then the interaction can have
both text parts and image parts. Each element is either a string or the
struct `llm-provider-utils-image` that I've added to
`llm-provider-utils.ell`.
4. `llm-provider-utils-image` has a both MIME type and a (binary) string
with the actual data.

While doing this, it made sense to add system prompts to vertex. I think
all their newer models support system prompts now, so hopefully this
won't cause any problems.

I think the same system should work for other providers, at least for
Claude, but I don't have any api keys of theirs to test out on.