Introduce llm-models-add, plus fix an issue with gemini 2.0 model (#136)
This is in response to a request from the discussion at
https://github.com/ahyatt/llm/discussions/134 by @s-kostyaev. The user
can add to the existing list of models via a function. It was possible
before as well, but this makes it more obvious, and adds a layer of
indirection in case we want to introduce a new variable for just
user-specified models.