Add the ability to return multiple outputs via a plist in llm calls (#160)
This adds the `multi-ouput` optional argument to the llm interface. When
set, the return value can contain different types of values: `text`, the
normal textual output of the model, `tool-results`, the results of tool
calls, `tool-uses` the uses determined by the LLM, from which we execute
to get the results, and `reasoning`, the output for reasoning.

This should enable us to solve https://github.com/ahyatt/llm/issues/143,
but it may require more work with the r1 server and any other reasoning
server. Other reasoning models didn't appear to have a different output
stream, but that may change.