Add reasoning control and support it for Claude and Ollama (#200)
Other tools like Open AI and Gemini also have reasoning controls, but
have not publicized the API for this, so they cannot yet support
reasoning.

This also adds support for getting Claude reasoning.

As part of this, we also increased the possible `max_tokens` for Claude,
which otherwise would have limited the reasoning output.

---------

Co-authored-by: Andrew Hyatt <ahyatt@continua.ai>